2025.04.15 플레이어 SRP 준수
PlayerController를 SRP원칙에 준수하여 수정

+++ Assets/Scripts/Interfaces/IDamageable.cs
@@ -1,4 +1,4 @@
 public interface IDamageable
 {
-    void GetDamage(int damage);
+    void GetDamage(float damage);
 }
